requirements.txt
flask
python-dotenv
gunicorn
requests